Add files via upload
This is a puzzle code in python language. This puzzle consists of a 5 X 5 frame which contained 24 small squared of equal size. A unique letter of the alphabet was printed on each small square. Since there were only 24 squares within the frame, the frame also contained an empty position which was the same size as a small square. A square could be moved into that empty position if it were immediately to the right, to the left, above, or below the empty position. The object of the puzzle was to slide squares into the empty position so that the frame displayed the alphabetical order.
--> The sequence of moves is represented by a sequence of As, Bs, Rs, and Ls to denote which square moves into the empty position.
--> A denotes that the square above the empty position moves
--> B denotes that the square below the empty position moves
--> L denotes that the square to the left of the empty positon moves
--> R denotes that the square to the right of the empty position moves