Add files via upload
A queue is another data structure, which is a collection of data items in which the data items are logically connected similar to a stack. But the insertion operation is performed only at one end, called 'rear', and the deletion operation is performed at the other end, called 'front'. 
This is a program to implement the queue operations(insertion, deletion) using C language.